Some refactoring to isolate ITs and STs as much as possible
