+++ test/test_helper.rb
@@ -14,6 +14,7 @@ require 'capybara/poltergeist'
 require 'capybara-screenshot/minitest'
 require 'omniauth_helper'
 require 'rack/test'
+require 'integration_system_test_helpers'
 Minitest::Reporters.use!
 
 Capybara.register_driver :poltergeist do |app|
@@ -33,68 +34,9 @@ end
 Capybara.save_path = "#{ENV.fetch('CIRCLE_ARTIFACTS', Rails.root.join('tmp/capybara'))}" if ENV['CIRCLE_ARTIFACTS']
 
 class ActionDispatch::IntegrationTest
-  include Capybara::DSL
-  include Capybara::Screenshot::MiniTestPlugin
-  include OmniauthMocker
-  OmniAuth.config.test_mode = true
+  include IntegrationSystemTestHelpers
 
   before { Capybara.reset_sessions! }
-
-  # for controllers
-  def sign_in(user)
-    post user_session_path \
-      "user[email]" => user.email,
-      "user[password]" => user.password
-  end
-
-  def choose_line(line)
-    post lines_path, params: { line: line.to_s }
-  end
-
-  # for proper integration tests
-  def log_in_as(user, line = 'DC')
-    log_in user
-    select_line line
-  end
-
-  def log_in(user)
-    visit root_path
-    fill_in 'Email', with: user.email
-    fill_in 'Password', with: user.password
-    click_button 'Sign in'
-  end
-
-  def select_line(line = 'DC')
-    choose line
-    click_button 'Select your line for this session'
-  end
-
-  def wait_for_element(text)
-    has_content? text
-  end
-
-  def wait_for_no_element(text)
-    has_no_content? text
-  end
-
-  def sign_out
-    click_link "#{@user.name}"
-    click_link 'Sign Out'
-  end
-
-  def wait_for_ajax
-    Timeout.timeout(Capybara.default_max_wait_time) do
-      loop until _finished_all_ajax_requests?
-    end
-  end
-
-  def _finished_all_ajax_requests?
-    page.evaluate_script('jQuery.active').zero?
-  end
-  
-  def go_to_dashboard
-    click_link "DARIA - #{(ENV['FUND'] ? ENV['FUND'] : Rails.env)}"
-  end
 end
 
 class ActionController::TestCase
